+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+- [Feature] Add isInline prop to Poptip component
 
 ## 40.3.0 - 2018-09-26
 - [Feature] Add isDisabled prop to Switch component